requirement.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
bootstrap4==0.1.0
boto3==1.13.17
botocore==1.16.17
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==14.0.0
django-bootstrap4==2.1.1
django-chartjs==2.2.1
django-crispy-forms==1.9.1
django-filter==2.3.0
django-storages==1.9.1
django-weasyprint==1.0.1
djangorestframework==3.12.1
docutils==0.15.2
gunicorn==20.0.4
heroku==0.1.4
html5lib==1.0.1
idna==2.9
jmespath==0.10.0
mysql-connector==2.2.9
Pillow==7.1.2
pycparser==2.20
PyMySQL==0.9.3
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.1.0
xlwt==1.3.0
